requirements.txt
# HAVEN Crowdfunding Backend - Production Dependencies
# Optimized for deployment with security and performance

# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
authlib==1.2.1
cryptography==41.0.8

# OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
facebook-sdk==3.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting and Security
slowapi==0.1.9

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Machine Learning (Lightweight)
scikit-learn==1.3.2
joblib==1.3.2

# Translation (Lightweight)
googletrans==4.0.0rc1

# Caching
redis==5.0.1

# File Processing
pillow==10.1.0
aiofiles==23.2.1

# Payment Integration
# Instamojo (using requests for API calls)

# Email Service Integration
sib-api-v3-sdk==7.6.0

# Search Service Integration
algoliasearch==3.0.0

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Monitoring and Logging
structlog==23.2.0

# Development and Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
secure==0.3.0

# Health Checks
fastapi-health==0.4.0

